+++ pyam/utils.py
@@ -20,11 +20,15 @@ except ImportError:
 logger = logging.getLogger(__name__)
 
 # common indices
+DEFAULT_META_INDEX = ['model', 'scenario']
 META_IDX = ['model', 'scenario']
 YEAR_IDX = ['model', 'scenario', 'region', 'year']
 IAMC_IDX = ['model', 'scenario', 'region', 'variable', 'unit']
 SORT_IDX = ['model', 'scenario', 'variable', 'year', 'region']
 LONG_IDX = IAMC_IDX + ['year']
+
+# required columns
+REQUIRED_COLS = ['region', 'variable', 'unit']
 
 # illegal terms for data/meta column names to prevent attribute conflicts
 ILLEGAL_COLS = ['data', 'meta']
@@ -73,6 +77,11 @@ def islistable(x):
     return isinstance(x, Iterable) and not isstr(x)
 
 
+def as_list(x):
+    """Return x as a list"""
+    return x if islistable(x) else [x]
+
+
 def write_sheet(writer, name, df, index=False):
     """Write a pandas.DataFrame to an ExcelWriter
 
@@ -107,33 +116,44 @@ def write_sheet(writer, name, df, index=False):
             pass
 
 
-def read_pandas(path, default_sheet='data', *args, **kwargs):
+def read_pandas(path, sheet_name='data*', *args, **kwargs):
     """Read a file and return a pandas.DataFrame"""
     if isinstance(path, Path) and path.suffix == '.csv':
-        df = pd.read_csv(path, *args, **kwargs)
+        return pd.read_csv(path, *args, **kwargs)
     else:
         xl = pd.ExcelFile(path)
-        if len(xl.sheet_names) > 1 and 'sheet_name' not in kwargs:
-            kwargs['sheet_name'] = default_sheet
-        df = pd.read_excel(path, *args, **kwargs)
+        sheet_names = pd.Series(xl.sheet_names)
 
-        # remove unnamed and empty columns
+        # reading multiple sheets
+        if len(sheet_names) > 1:
+            sheets = kwargs.pop('sheet_name', sheet_name)
+            # apply pattern-matching for sheet names (use * as wildcard)
+            sheets = sheet_names[pattern_match(sheet_names, values=sheets)]
+            if sheets.empty:
+                raise ValueError(f'No sheets {sheet_name} in file {path}!')
+
+            df = pd.concat([xl.parse(s, *args, **kwargs) for s in sheets])
+
+        # read single sheet (if only one exists in file) ignoring sheet name
+        else:
+            df = pd.read_excel(path, *args, **kwargs)
+
+        # remove unnamed and empty columns, and rows were all values are nan
         empty_cols = [c for c in df.columns if str(c).startswith('Unnamed: ')
                       and all(np.isnan(df[c]))]
-        df.drop(columns=empty_cols, inplace=True)
-    return df
+        return df.drop(columns=empty_cols).dropna(axis=0, how='all')
 
 
 def read_file(path, *args, **kwargs):
     """Read data from a file"""
-    format_kwargs = {}
     # extract kwargs that are intended for `format_data`
+    format_kwargs = dict(index=kwargs.pop('index'))
     for c in [i for i in IAMC_IDX + ['year', 'time', 'value'] if i in kwargs]:
         format_kwargs[c] = kwargs.pop(c)
     return format_data(read_pandas(path, *args, **kwargs), **format_kwargs)
 
 
-def format_data(df, **kwargs):
+def format_data(df, index, **kwargs):
     """Convert a pandas.Dataframe or pandas.Series to the required format"""
     if isinstance(df, pd.Series):
         df.name = df.name or 'value'
@@ -214,29 +234,33 @@ def format_data(df, **kwargs):
     if conflict_cols:
         msg = f'Column name {conflict_cols} is illegal for timeseries data.\n'
         _args = ', '.join([f"{i}_1='{i}'" for i in conflict_cols])
-        msg += f'Use `IamDataFrame(..., {_args})` to rename at initalization.'
+        msg += f'Use `IamDataFrame(..., {_args})` to rename at initialization.'
         raise ValueError(msg)
 
-    # format columns to lower-case and check that all required columns exist
-    if not set(IAMC_IDX).issubset(set(df.columns)):
-        missing = list(set(IAMC_IDX) - set(df.columns))
-        raise ValueError("missing required columns `{}`!".format(missing))
+    # check that index and required columns exist
+    missing_index = [c for c in index if c not in df.columns]
+    if missing_index:
+        raise ValueError(f'Missing index columns `{missing_index}`!')
+
+    missing_required_col = [c for c in REQUIRED_COLS if c not in df.columns]
+    if missing_required_col:
+        raise ValueError(f'Missing required columns `{missing_required_col}`!')
 
     # check whether data in wide format (IAMC) or long format (`value` column)
     if 'value' in df.columns:
         # check if time column is given as `year` (int) or `time` (datetime)
-        cols = df.columns
-        if 'year' in cols:
+        if 'year' in df.columns:
             time_col = 'year'
-        elif 'time' in cols:
+        elif 'time' in df.columns:
             time_col = 'time'
         else:
-            msg = 'invalid time format, must have either `year` or `time`!'
+            msg = 'Invalid time format, must have either `year` or `time`!'
             raise ValueError(msg)
-        extra_cols = list(set(cols) - set(IAMC_IDX + [time_col, 'value']))
+        extra_cols = [c for c in df.columns
+                      if c not in index + REQUIRED_COLS + [time_col, 'value']]
     else:
         # if in wide format, check if columns are years (int) or datetime
-        cols = set(df.columns) - set(IAMC_IDX)
+        cols = [c for c in df.columns if c not in index + REQUIRED_COLS]
         year_cols, time_cols, extra_cols = [], [], []
         for i in cols:
             try:
@@ -257,7 +281,8 @@ def format_data(df, **kwargs):
         else:
             msg = 'invalid column format, must be either years or `datetime`!'
             raise ValueError(msg)
-        df = pd.melt(df, id_vars=IAMC_IDX + extra_cols, var_name=time_col,
+        cols = index + REQUIRED_COLS + extra_cols
+        df = pd.melt(df, id_vars=cols, var_name=time_col,
                      value_vars=sorted(melt_cols), value_name='value')
 
     # cast value column to numeric and drop nan
@@ -273,7 +298,7 @@ def format_data(df, **kwargs):
         _raise_data_error('empty cells in `data`', df.loc[null_rows])
 
     # check for duplicates and empty data
-    idx_cols = IAMC_IDX + [time_col] + extra_cols
+    idx_cols = index + REQUIRED_COLS + [time_col] + extra_cols
     rows = df[idx_cols].duplicated()
     if any(rows):
         _raise_data_error('duplicate rows in `data`', df.loc[rows, idx_cols])
@@ -282,7 +307,7 @@ def format_data(df, **kwargs):
         logger.warning('Formatted data is empty!')
 
     df = format_time_col(sort_data(df, idx_cols), time_col)
-    return df, time_col, extra_cols
+    return df.set_index(idx_cols).value, index, time_col, extra_cols
 
 
 def format_time_col(data, time_col):
@@ -307,7 +332,7 @@ def sort_data(data, cols):
     return data.sort_values(cols)[cols + ['value']].reset_index(drop=True)
 
 
-def merge_meta(left, right, ignore_meta_conflict=False):
+def merge_meta(left, right, ignore_conflict=False):
     """Merge two `meta` tables; raise if values are in conflict (optional)
 
     If conflicts are ignored, values in `left` take precedence over `right`.
@@ -319,7 +344,7 @@ def merge_meta(left, right, ignore_meta_conflict=False):
     # merge `right` into `left` for overlapping scenarios ( `sect`)
     if not sect.empty:
         # if not ignored, check that overlapping `meta` columns are equal
-        if not ignore_meta_conflict:
+        if not ignore_conflict:
             cols = [i for i in right.columns if i in left.columns]
             if not left.loc[sect, cols].equals(right.loc[sect, cols]):
                 conflict_idx = (
@@ -405,8 +430,7 @@ def pattern_match(data, values, level=None, regexp=False, has_nan=False):
     for filtering (str, int, bool)
     """
     matches = np.array([False] * len(data))
-    if not isinstance(values, Iterable) or isstr(values):
-        values = [values]
+    values = values if islistable(values) else [values]
 
     # issue (#40) with string-to-nan comparison, replace nan by empty string
     _data = data.copy()
@@ -614,3 +638,8 @@ def get_variable_components(x, level, join=False):
         level = [level] if type(level) == int else level
         join = '|' if join is True else join
         return join.join([_x[i] for i in level])
+
+
+def s(n):
+    """Return an s if n!=1 for nicer formatting of log messages"""
+    return 's' if n != 1 else ''